requirements.txt
langchain
langchain-huggingface
huggingface-hub
sentence-transformers
faiss-cpu
python-dotenv
streamlit
langchain-community